Prevent integer overflow when dumping.
Fixes a1820573705d9e105cb5b46e44db639199e02b03

+++ cli/cli_dump.cpp
@@ -273,7 +273,7 @@ command(int argc, char *argv[])
         while ((call = p.parse_call())) {
             // Give a few calls of tolerance before bailing out to allow pending
             // multi-threaded calls out of order to dump
-            const unsigned call_no_tol = 100;
+            const unsigned long long call_no_tol = 100;
 
             if (call->no > calls.getLast() + call_no_tol) {
                 delete call;
